Update and rename Info/sistema_operacional.py to classes/sistema_operacional.py
In the update, were removed the reading of existing users from a .txt file, because it wasn't supposed to work that way. For it to work, the users should have been saved in a file (inside the blocks of an inode) created using the emulated file system and saved in the virtual disc, which is the only .txt file allowed in the project.

Also, the "criar_usuario" (create user) has been corrected. It was only checking if the username existed and if it did, it would raise an exception, and if it didn't exist, the user wasn't being created. Now it is.

Beside the corrections, a new function "converter_caminho_para_lista" (convert path to list) has been created. It is supposed to convert a string path to a list with the separated paths. For example: "/home/my_folder/my_other_folder/" -> ['home', 'my_folder', 'my_other_folder']. It still has some problems to be corrected, nonetheless it is working as expected.