Add files via upload
Added the "bloco" (block)  and the "sistema_arquivos" (file system) class files. To-do: check side-by-side if there's something that can be used from the "/gerenciador-arquivos/Info/sistema_arquivos.py", which was developed at the same time as this one. But, by doing a quick analysis, there has been noted some mistakes made in that one, which will need to be redone in order to make it work as expected.